+++ kaithem/src/systasks.py
@@ -10,7 +10,6 @@ import threading
 import time
 
 import quart
-import scullery.workers
 import structlog
 from scullery import scheduling
 from zeroconf import ServiceBrowser, ServiceStateChange
@@ -210,13 +209,6 @@ def logstats():
                 MemUseWasTooHigh = False
         except Exception as e:
             raise e
-
-
-def stop_workers(*a):
-    scullery.workers.stop()
-
-
-messagebus.subscribe("/system/shutdown", stop_workers)
 
 
 def sd():